% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akl_transport.R
\name{get_akl_calendar_dates}
\alias{get_akl_calendar_dates}
\title{Get calendar dates}
\usage{
get_akl_calendar_dates()
}
\value{
A tibble
}
\description{
Query calendar dates
}
\examples{
get_akl_calendar_dates()
}
