% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akl_transport.R
\name{get_calender_by_service}
\alias{get_calender_by_service}
\title{Get calendar by service id}
\usage{
get_calender_by_service(service_id = NULL)
}
\arguments{
\item{service_id}{string service id}
}
\value{
A tibble
}
\description{
Query calendar by service id
}
\examples{
\dontrun{
get_calender_by_service(service_id = "467236433-20201222141234_v95.107")
}
# get calender by service
}
