% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akl_transport.R
\name{get_info_by_stop_code}
\alias{get_info_by_stop_code}
\title{Get stop info by stop code}
\usage{
get_info_by_stop_code(stop_code = NULL)
}
\arguments{
\item{stop_code}{string code of the stop}
}
\value{
A tibble
}
\description{
Query stop info by stop code
}
\examples{
get_info_by_stop_code(stop_code = "1371")
# get info by stop_code
}
