% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akl_transport.R
\name{get_routes_by_location}
\alias{get_routes_by_location}
\title{Get routes by location}
\usage{
get_routes_by_location(lat = NULL, lng = NULL, distance = NULL)
}
\arguments{
\item{lat}{string latitude of the position}

\item{lng}{string longitude of the position}

\item{distance}{string search radius}
}
\value{
A tibble
}
\description{
Query routes by latitude of the position,
longitude of the position,and search radius
}
\examples{
get_routes_by_location(
  lat = "-37.06429",
  lng = "174.94611",
  distance = "10"
)
# get routes by location
}
