% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akl_transport.R
\name{get_routes_by_long_name}
\alias{get_routes_by_long_name}
\title{Get routes by long name}
\usage{
get_routes_by_long_name(route_long_name = NULL)
}
\arguments{
\item{route_long_name}{string name of the rout}
}
\value{
A tibble
}
\description{
Query routes by long name
}
\examples{
get_routes_by_long_name(route_long_name = "Mangere Town Centre To Otahuhu Station Via Tidal Road")
}
