% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akl_transport.R
\name{get_routes_by_route_id}
\alias{get_routes_by_route_id}
\title{Get routes by routes id}
\usage{
get_routes_by_route_id(route_id = NULL)
}
\arguments{
\item{route_id}{string identifier of the route}
}
\value{
A tibble
}
\description{
Query routes by routes id
}
\examples{
get_routes_by_route_id(route_id = "84303-20201221160823_v95.106")
# get routes by route_id
}
