% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akl_transport.R
\name{get_routes_by_search_string_and_route_types}
\alias{get_routes_by_search_string_and_route_types}
\title{Get routes by search_string and route_types}
\usage{
get_routes_by_search_string_and_route_types(
  search_string = NULL,
  route_types = NULL
)
}
\arguments{
\item{search_string}{string}

\item{route_types}{string}
}
\value{
A tibble
}
\description{
Query routes by search_string and route_types
}
\examples{
get_routes_by_search_string_and_route_types(search_string = "Albany Station",
                                            route_types = c("3","2"))
}
