% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akl_transport.R
\name{get_routes_by_search_text}
\alias{get_routes_by_search_text}
\title{Get routes by search text}
\usage{
get_routes_by_search_text(search_text = NULL)
}
\arguments{
\item{search_text}{string text to search}
}
\value{
A tibble
}
\description{
Query routes by search text
}
\examples{
get_routes_by_search_text(search_text = "Albany")
}
