% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akl_transport.R
\name{get_routes_by_short_name}
\alias{get_routes_by_short_name}
\title{Get routes by short name}
\usage{
get_routes_by_short_name(route_short_name = NULL)
}
\arguments{
\item{route_short_name}{string name of the route}
}
\value{
A tibble
}
\description{
Query routes by short name
}
\examples{
get_routes_by_short_name(route_short_name = "326")
}
