% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akl_transport.R
\name{get_routes_by_stop}
\alias{get_routes_by_stop}
\title{Get routes by stop}
\usage{
get_routes_by_stop(stop_id = NULL)
}
\arguments{
\item{stop_id}{string stop to search}
}
\value{
A tibble
}
\description{
Query routes by stop id
}
\examples{
\dontrun{
get_routes_by_stop(stop_id = "1369-20201221160823_v95.106")
}
# get routes by stop
}
