% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akl_transport.R
\name{get_shape_geometry_by_shape_id}
\alias{get_shape_geometry_by_shape_id}
\title{Get shape geometry by shape id}
\usage{
get_shape_geometry_by_shape_id(shape_id = NULL)
}
\arguments{
\item{shape_id}{string identifier of the shape}
}
\value{
A tibble
}
\description{
Query shape geometry by shape id
}
\examples{
\dontrun{
get_shape_geometry_by_shape_id(shape_id = "1192-20201221160823_v95.106")
}
# get shape geometry by shape_id
}
