% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akl_transport.R
\name{get_shapes_by_shape_id}
\alias{get_shapes_by_shape_id}
\title{Get shapes by shape id}
\usage{
get_shapes_by_shape_id(shape_id = NULL)
}
\arguments{
\item{shape_id}{string identifier of the shape}
}
\value{
A tibble
}
\description{
Query shapes by shape id
}
\examples{
get_shapes_by_shape_id(shape_id = "1192-20201221160823_v95.106")
# get shapes by shape_id
}
