% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akl_transport.R
\name{get_stop_by_stop_code}
\alias{get_stop_by_stop_code}
\title{Get stop by stop code}
\usage{
get_stop_by_stop_code(stop_code = NULL)
}
\arguments{
\item{stop_code}{string code of the stop}
}
\value{
A tibble
}
\description{
Query stop by stop code
}
\examples{
get_stop_by_stop_code(stop_code = "97")
# get stop by stop_code
}
