% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akl_transport.R
\name{get_stop_by_stop_id}
\alias{get_stop_by_stop_id}
\title{Get stop by stop id}
\usage{
get_stop_by_stop_id(stop_id = NULL)
}
\arguments{
\item{stop_id}{string identifier of the stop}
}
\value{
A tibble
}
\description{
Query stop by stop id
}
\examples{
\dontrun{
get_stop_by_stop_id(stop_id = "1369-20201221160823_v95.106")
}
# get stop by stop_id
}
