% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akl_transport.R
\name{get_stop_by_trip_from_stop}
\alias{get_stop_by_trip_from_stop}
\title{Get stop by Trip from stop}
\usage{
get_stop_by_trip_from_stop(trip_id = NULL, stop_id = NULL)
}
\arguments{
\item{trip_id}{string trip unique identifier}

\item{stop_id}{string stop unique identifier}
}
\value{
A tibble
}
\description{
Query stop by Trip from stop
}
\examples{
get_stop_by_trip_from_stop(
  trip_id = "51100221657-20201221160823_v95.106",
  stop_id = "0140-20201221160823_v95.106"
)
# get stop by trip from stop
}
