% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akl_transport.R
\name{get_stop_by_trip_id}
\alias{get_stop_by_trip_id}
\title{Get stop by trip id}
\usage{
get_stop_by_trip_id(trip_id = NULL)
}
\arguments{
\item{trip_id}{string trip unique identifier}
}
\value{
A tibble
}
\description{
Query stop by trip id
}
\examples{
get_stop_by_trip_id(trip_id = "51100221657-20201221160823_v95.106")
}
