% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akl_transport.R
\name{get_stop_times_by_stop_id}
\alias{get_stop_times_by_stop_id}
\title{Get stop times by stop id}
\usage{
get_stop_times_by_stop_id(stop_id = NULL)
}
\arguments{
\item{stop_id}{string stop id}
}
\value{
A tibble
}
\description{
Query stop times by stop id
}
\examples{
get_stop_times_by_stop_id(stop_id = "1369-20201221160823_v95.106")
}
