% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akl_transport.R
\name{get_stop_times_by_trip_and_sequence}
\alias{get_stop_times_by_trip_and_sequence}
\title{Get stop times by trip and sequence}
\usage{
get_stop_times_by_trip_and_sequence(trip_id = NULL, stop_sequence = NULL)
}
\arguments{
\item{trip_id}{string trip id}

\item{stop_sequence}{integer Format - int64. stop sequence}
}
\value{
A tibble
}
\description{
Query stop times by trip id and sequence
}
\examples{
get_stop_times_by_trip_and_sequence(trip_id = "465203730-20201221160823_v95.106",
                                    stop_sequence = "1")
}
