% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akl_transport.R
\name{get_stop_times_by_trip_id}
\alias{get_stop_times_by_trip_id}
\title{Get stop times by trip id}
\usage{
get_stop_times_by_trip_id(trip_id = NULL)
}
\arguments{
\item{trip_id}{string trip id}
}
\value{
A tibble
}
\description{
Query stop times by trip id
}
\examples{
get_stop_times_by_trip_id(trip_id = "465203730-20201221160823_v95.106")
}
