% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akl_transport.R
\name{get_stops_by_location}
\alias{get_stops_by_location}
\title{Get stops by location}
\usage{
get_stops_by_location(lat = NULL, lng = NULL, distance = NULL)
}
\arguments{
\item{lat}{string latitude of the position}

\item{lng}{string longitude of the position}

\item{distance}{string search radiu}
}
\value{
A tibble
}
\description{
Query stops by location
}
\examples{
get_stops_by_location(lat = "-37.06429",
                      lng = "174.94611",
                      distance = "10")
}
