% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akl_transport.R
\name{get_stops_by_route_ids}
\alias{get_stops_by_route_ids}
\title{Get stops by route ids}
\usage{
get_stops_by_route_ids(route_ids = NULL)
}
\arguments{
\item{route_ids}{string comma separated list of route ids}
}
\value{
A tibble
}
\description{
Query stops by route ids
}
\examples{
get_stops_by_route_ids(route_ids = c("06259-20201126121446_v95.73,32601-20201205123725_v95.82"))
# get stops through BTF search(route_ids)
}
