% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akl_transport.R
\name{get_stops_by_search_text}
\alias{get_stops_by_search_text}
\title{Get stops by search text}
\usage{
get_stops_by_search_text(search_text = NULL)
}
\arguments{
\item{search_text}{string search text to find stops}
}
\value{
A tibble
}
\description{
Query stops by search text
}
\examples{
\dontrun{
get_stops_by_search_text(search_text = "Hinemoa St")
}
# get stops by search_text
}
