% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akl_transport.R
\name{get_timetable_by_route_ids_stop_code_and_short_name}
\alias{get_timetable_by_route_ids_stop_code_and_short_name}
\title{Get timetable by short_name,route_ids and stop_code}
\usage{
get_timetable_by_route_ids_stop_code_and_short_name(
  route_ids = NULL,
  stop_code = NULL,
  route_short_name = NULL
)
}
\arguments{
\item{route_ids}{string comma separated route ids}

\item{stop_code}{string stop code}

\item{route_short_name}{string route short name like (STH, 923)}
}
\value{
A tibble
}
\description{
Query timetable
}
\examples{
get_timetable_by_route_ids_stop_code_and_short_name(route_ids = c("90111-20201205123725_v95.82,05363-20201205123725_v95.82"),
                                                                  stop_code = "3637",
                                                                  route_short_name = "901")
}
