% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akl_transport.R
\name{get_trips_by_route}
\alias{get_trips_by_route}
\title{Get trips by route}
\usage{
get_trips_by_route(route_id = NULL)
}
\arguments{
\item{route_id}{string route id}
}
\value{
A tibble
}
\description{
Query trips by route id
}
\examples{
get_trips_by_route(route_id = "840007-20201221160823_v95.106")
}
