% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akl_transport.R
\name{get_trips_by_trip_id}
\alias{get_trips_by_trip_id}
\title{Get trips by trip}
\usage{
get_trips_by_trip_id(trip_id = NULL)
}
\arguments{
\item{trip_id}{string trip id}
}
\value{
A tibble
}
\description{
Query trips by trip id
}
\examples{
get_trips_by_trip_id(trip_id = "465203730-20201221160823_v95.106")
# get trips by trip_id
}
