% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kiaora-package.R
\docType{package}
\name{kiaora-package}
\alias{kiaora}
\alias{kiaora-package}
\title{kiaora}
\description{
This is a package with functions to get Auckland transport 
    information and residential auction results for NZ property, including 
    property information,auction price and dates, rating value and dates.
}
\author{
\strong{Maintainer}: Tina Ye \email{xye112@aucklanduni.ac.nz}

}
