% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzhousingprice.R
\docType{data}
\name{nzpropertygeo}
\alias{nzpropertygeo}
\title{New Zealand property address}
\format{
A \code{tibble} with 8 columns, which are:
\describe{
\item{property_address}{address for auction property}
\item{auction_price}{price for auction property}
\item{auction_dates}{auction dates for property}
\item{bedrooms}{number of bedrooms for property}
\item{bathrooms}{number of bathrooms for property}
\item{car_parking}{number of garaging for property}
\item{rating_value}{rating value for property}
\item{rating_dates}{rating dates for property}
}
}
\usage{
nzpropertygeo
}
\description{
A tibble that contains NZ property address and geographical locations
}
\keyword{datasets}
