% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akl_transport.R
\name{get_akl_agency}
\alias{get_akl_agency}
\title{Get Auckland bus agency}
\usage{
get_akl_agency(api_key = NULL)
}
\arguments{
\item{api_key}{A string that contains your API key.}
}
\value{
A tibble
}
\description{
Query Auckland bus agency
}
