% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kiaora-package.R
\docType{package}
\name{kiaora-package}
\alias{kiaora}
\alias{kiaora-package}
\title{kiaora}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Tina Ye \email{xye112@aucklanduni.ac.nz}

}
